Documentation in components, styling adjusted
