+++ client/src/styles/pages/MainPage.css
@@ -9,7 +9,3 @@
 .content-container img {
 	margin-right: 20px; /* Add some space between the gif and the Community component */
 }
-
-.text-container {
-	background-color: rgb(23, 78, 209);
-}
